feat: Add comprehensive features for recruitment, performance, payroll loans, and leave management.

+++ Backend/HRMS/HRMS.Application/Features/Payroll/Loans/Queries/GetMonthlyInstallments/GetMonthlyInstallmentsQuery.cs
@@ -42,7 +42,7 @@ public class GetMonthlyInstallmentsQueryHandler : IRequestHandler<GetMonthlyInst
 
         var dtos = installments.Select(i => new LoanInstallmentDto
         {
-            InstallmentId = i.InstallmentId,
+            InstallmentId = (int)i.InstallmentId,
             LoanId = i.LoanId,
             EmployeeId = i.Loan.EmployeeId,
             EmployeeName = i.Loan.Employee.FullNameAr,
